move job helpers away
